+++ src/import.ts
@@ -1,14 +1,21 @@
-import * as admin from "firebase-admin";
-import * as fs from "fs";
+import * as fs from 'fs';
+import { v1 as uuidv1 } from 'uuid';
+import * as admin from 'firebase-admin';
+
+export interface IImportOptions {
+  dates?: string[];
+  geos?: string[];
+  refs?: string[];
+}
 
 /**
  * Convert time array in document object
- * @param dateArray 
- * @param documentObj 
+ * @param dateArray
+ * @param documentObj
  */
 const updateTime = (dateArray, documentObj): void => {
   dateArray.forEach(c => {
-    c.split(".").reduce((acc, cur, i, a) => {
+    c.split('.').reduce((acc, cur, i, a) => {
       if (!a[i + 1] && acc[cur] && acc[cur]._seconds) {
         acc[cur] = new Date(acc[cur]._seconds * 1000);
       } else return acc[cur] || {};
@@ -20,32 +27,30 @@ const updateTime = (dateArray, documentObj): void => {
  * Restore data to firestore
  *
  * @param {string} fileName
- * @param {Array<string>} dateArray
- * @param {Array<string>} geoArray
+ * @param {IImportOptions} options
  */
 export const restore = (
   fileName: string,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions
 ): Promise<any> => {
   const db = admin.firestore();
 
   return new Promise((resolve, reject) => {
-    if (typeof fileName === "object") {
+    if (typeof fileName === 'object') {
       let dataArray = fileName;
 
-      updateCollection(db, dataArray, dateArray, geoArray)
+      updateCollection(db, dataArray, options)
         .then(() => {
           resolve({
             status: true,
-            message: "Collection successfully imported!"
+            message: 'Collection successfully imported!'
           });
         })
         .catch(error => {
           reject({ status: false, message: error.message });
         });
     } else {
-      fs.readFile(fileName, "utf8", function(err, data) {
+      fs.readFile(fileName, 'utf8', function(err, data) {
         if (err) {
           console.log(err);
           reject({ status: false, message: err.message });
@@ -54,11 +59,11 @@ export const restore = (
         // Turn string from file to an Array
         let dataArray = JSON.parse(data);
 
-        updateCollection(db, dataArray, dateArray, geoArray)
+        updateCollection(db, dataArray, options)
           .then(() => {
             resolve({
               status: true,
-              message: "Collection successfully imported!"
+              message: 'Collection successfully imported!'
             });
           })
           .catch(error => {
@@ -66,7 +71,7 @@ export const restore = (
           });
       });
     }
-  });
+  }).catch(error => console.error(error));
 };
 
 /**
@@ -80,33 +85,33 @@ export const restore = (
 const updateCollection = async (
   db,
   dataArray: Array<any>,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions = {}
 ) => {
   for (var index in dataArray) {
     var collectionName = index;
     for (var doc in dataArray[index]) {
       if (dataArray[index].hasOwnProperty(doc)) {
-        if (dataArray[index][doc]["subCollection"]) {
-          const subCollections = dataArray[index][doc]["subCollection"];
-          delete dataArray[index][doc]["subCollection"];
+        // asign document id for array type
+        let docId = Array.isArray(dataArray[index]) ? uuidv1() : doc;
+        if (dataArray[index][doc]['subCollection']) {
+          const subCollections = dataArray[index][docId]['subCollection'];
+          delete dataArray[index][doc]['subCollection'];
           await startUpdating(
             db,
             collectionName,
-            doc,
+            docId,
             dataArray[index][doc],
-            dateArray,
-            geoArray
+            options
           );
-          await updateCollection(db, subCollections, [], []);
+          // Update sub collection
+          await updateCollection(db, subCollections);
         } else {
           await startUpdating(
             db,
             collectionName,
-            doc,
+            docId,
             dataArray[index][doc],
-            dateArray,
-            geoArray
+            options
           );
         }
       }
@@ -118,57 +123,64 @@ const updateCollection = async (
  * Write data to database
  * @param db
  * @param collectionName
- * @param doc
+ * @param docId
  * @param data
  * @param dateArray
  * @param geoArray
  */
+
 const startUpdating = (
   db,
   collectionName: string,
-  doc: string,
+  docId: string,
   data: object,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions
 ) => {
-  let parameterValid = true;
+  // Update date value
+  if (options.dates && options.dates.length > 0) {
+    updateTime(options.dates, data);
+  }
 
-  if (typeof dateArray === "object" && dateArray.length > 0) {
-    updateTime(dateArray, data);    
+  // reference key
+  if (options.refs && options.refs.length > 0) {
+    options.refs.forEach(ref => {
+      if (data.hasOwnProperty(ref)) {
+        data[ref] = db.doc(data[ref]);
+      }
+    });
   }
 
   // Enter geo value
-  if (typeof geoArray !== "undefined" && geoArray.length > 0) {
-    geoArray.forEach(geo => {
+  if (options.geos && options.geos.length > 0) {
+    options.geos.forEach(geo => {
       if (data.hasOwnProperty(geo)) {
         data[geo] = new admin.firestore.GeoPoint(
           data[geo]._latitude,
           data[geo]._longitude
         );
       } else {
-        console.log("Please check your geo parameters!!!", geoArray);
-        parameterValid = false;
+        console.warn('Please check your geo parameters!!!', options.geos);
       }
     });
   }
 
-  if (parameterValid) {
-    return new Promise(resolve => {
-      db.collection(collectionName)
-        .doc(doc)
-        .set(data)
-        .then(() => {
-          console.log(`${doc} was successfully added to firestore!`);
-          resolve("Data written!");
-        })
-        .catch(error => {
-          console.log(error);
+  return new Promise((resolve, reject) => {
+    db.collection(collectionName)
+      .doc(docId)
+      .set(data)
+      .then(() => {
+        console.log(`${docId} was successfully added to firestore!`);
+        resolve({
+          status: true,
+          message: `${docId} was successfully added to firestore!`
         });
-    });
-  } else {
-    console.log(
-      `${doc} was not imported to firestore. Please check your parameters or ignore if you don't need to import the property above.`
-    );
-    return false;
-  }
+      })
+      .catch(error => {
+        console.log(error);
+        reject({
+          status: false,
+          message: error.message
+        });
+      });
+  });
 };