+++ CHANGELOG.md
@@ -11,6 +11,22 @@
 
 ---
 
+## [0.3.3] - 2020-01-03
+
+#### - :nail_care: [Polish]
+
+- Update new packages
+- Check app before initializing
+
+#### - :boom: [Breaking Change]
+
+- Changed options for date & location & ref
+
+#### - :rocket: [New Feature]
+
+- Added support reference type
+- Added upload array of document feature (without predefined document id)
+
 ## [0.3.2] - 2020-01-03
 
 #### - :rocket: [New Feature]